+++ processor/src/test/resources/input/com/example/arez/AllowNoArezDepsComponent.java
@@ -1,0 +1,18 @@
+package com.example.arez;
+
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.ReactComponent;
+import react4j.arez.ReactArezComponent;
+
+@ReactComponent( allowNoArezDeps = true )
+abstract class AllowNoArezDepsComponent
+  extends ReactArezComponent
+{
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}